requirements.txt
Django==2.2.4
sphinx==2.2.0
django-docs
bs4
djangorestframework
djangorestframework-jwt
django-cors-headers
django-redis-sessions==0.6.1
django-redis
tornado==6.0.3
tornado-redis
python-socketio
eventlet
aiohttp
aioredis
requests
Pillow
django-tinymce
django-filebrowser
django-grappelli==2.12.3
django-image-cropping
easy-thumbnails
celery
django-celery-results
django-celery-beat
django-cleanup
django-rosetta 
django-modeltranslation==0.13.3
grappelli-modeltranslation==0.1.3
django-countries
django-grappelli-extras==1.2.0 
sphinxcontrib-plantuml==0.17.1
sphinx-autobuild